@ php_yod

+++ yod/yod_application.c
@@ -421,7 +421,7 @@ int yod_application_config(char *name, uint name_len, zval *result TSRMLS_DC) {
 int yod_application_import(char *alias, uint alias_len, char *classext, uint classext_len TSRMLS_DC) {
 	zval *p_imports, **ppval;
 	char *classfile, *classname, *classpath;
-	uint classfile_len, classname_len;
+	size_t classfile_len, classname_len;
 	zend_class_entry **pce = NULL;
 
 #if PHP_YOD_DEBUG
